MB-67520: Enable narrowing compile errors for Clang/MSVC
Enable:

For clang
-Werror=shorten-64-to-32 -Werror=implicit-int-conversion

For MSVC
/W4

For a selected number of valuable targets which are now clean
of warnings.

Reviewed-on: https://review.couchbase.org/c/kv_engine/+/231309
Tested-by: Build Bot <build@couchbase.com>

+++ daemon/connection.cc
@@ -1124,10 +1124,11 @@ void Connection::setAuthenticated(cb::rbac::UserIdent ui) {
 #ifdef __linux__
         if (!listening_port->system) {
             auto timeout = Settings::instance().getTcpUserTimeout();
-            if (!cb::net::setSocketOption<uint32_t>(socketDescriptor,
-                                                    IPPROTO_TCP,
-                                                    TCP_USER_TIMEOUT,
-                                                    timeout.count())) {
+            if (!cb::net::setSocketOption<uint32_t>(
+                        socketDescriptor,
+                        IPPROTO_TCP,
+                        TCP_USER_TIMEOUT,
+                        gsl::narrow_cast<uint32_t>(timeout.count()))) {
                 LOG_WARNING_CTX(
                         "Failed to set TCP_USER_TIMEOUT",
                         {"conn_id", getId()},
